выполнил задание модульность
