+++ js/constants.js
@@ -1,0 +1,1 @@
+export const SIMILIAR_POSTS_COUNT = 25;
