+++ js/data.js
@@ -1,0 +1,18 @@
+import {
+  SIMILIAR_POSTS_COUNT,
+} from './constants.js';
+
+import {
+  createUniqValues,
+} from './utils.js';
+
+export const POST_DESCRIPTIONS = [
+  'Далеко-далеко за, словесными горами в стране гласных и согласных живут рыбные тексты. Его если путь родного вдали переписали но своих реторический снова, текста языком рыбного журчит языкового.',
+  'Которой рукопись то букв скатился всеми текстов пояс свой, заголовок она. Букв залетают текстами переулка моей большого курсивных не последний запятых переписали встретил снова океана всемогущая даль коварных.',
+  'Продолжил необходимыми ему текстами живет запятой текста инициал сбить. Толку, лучше скатился. Щеке до образ страна бросил строчка вершину реторический!',
+  'Его подзаголовок даль журчит реторический. Свое текстов путь большой пор над повстречался вершину, там решила! Гор все своего, рот сбить они образ.',
+  'Даль ручеек коварный точках реторический! Возвращайся домах решила имеет дал языком дорогу, не последний сих! Собрал до все что заголовок, строчка ведущими злых рукописи о, ручеек семантика, составитель коварных это?',
+  'Послушавшись, которой бросил она продолжил заголовок которое текстов использовало? Жаренные взобравшись продолжил точках необходимыми строчка, подпоясал,  своих сих щеке путь! Речью однажды грамматики запятой!'
+];
+export const uniqueIDs = createUniqValues( SIMILIAR_POSTS_COUNT );
+export const uniqueFileNames = createUniqValues( SIMILIAR_POSTS_COUNT );
